Devex 390: Consolidate shared utilities folders
