+++ web-api/src/getDynamoEndpoints.js
