+++ web-api/src/getDynamoEndpoints.test.js
